# internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suxen control-plane API",
    "version": "dev",
    "description": "Administrative API for repositories, access control, retention, provenance, and operations. OCI Distribution endpoints under /v2 and /repository/{name}/v2 use the OCI error envelope and are intentionally outside this control-plane contract."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current suxen server"
    }
  ],
  "security": [
    {},
    {
      "basicAuth": []
    },
    {
      "bearerAuth": []
    },
    {
      "oidcSession": []
    }
  ],
...
  "paths": {
...
    "/api/v1/repositories/{name}/download-gate": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RepositoryName"
        }
      ],
      "get": {
        "operationId": "getDownloadGate",
        "x-suxen-required-privilege": "repository:{name}:read",
        "summary": "Get the repository download gate",
        "tags": [
          "Policies"
        ],
        "responses": {
          "200": {
            "description": "Download gate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadGate"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "setDownloadGate",
        "x-suxen-required-privilege": "repository:{name}:manage",
        "parameters": [{"$ref": "#/components/parameters/Force"}],
        "summary": "Set the repository download gate",
        "description": "Only hosted and proxy repositories support download gates. Downloads through a group enforce the supplying member's gate.",
        "tags": [
          "Policies"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadGateUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Download gate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadGate"
                }
              }
            }
          },
          "201": {
            "description": "Download gate created",
            "headers": {"Location": {"$ref": "#/components/headers/Location"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DownloadGate"}}}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteDownloadGate",
        "x-suxen-required-privilege": "repository:{name}:manage",
        "parameters": [{"$ref": "#/components/parameters/Force"}],
        "summary": "Delete the repository download gate",
        "tags": [
          "Policies"
        ],
        "responses": {
          "204": {
            "description": "Download gate deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {"$ref": "#/components/responses/Conflict"},
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/download-gate-defaults": {
      "get": {
        "operationId": "getDownloadGateDefaults",
        "x-suxen-required-privilege": "admin:download-gate-defaults:read",
        "summary": "Get the instance-wide download-gate default",
        "tags": [
          "Policies"
        ],
        "responses": {
          "200": {
            "description": "Instance-wide download-gate default",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadGate"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "setDownloadGateDefaults",
        "x-suxen-required-privilege": "admin:download-gate-defaults:write",
        "summary": "Set the instance-wide download-gate default",
        "tags": [
          "Policies"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadGateDefaultsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instance-wide download-gate default",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadGate"
                }
              }
            }
          },
          "201": {
            "description": "Instance-wide download-gate default created",
            "headers": {"Location": {"$ref": "#/components/headers/Location"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DownloadGate"}}}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteDownloadGateDefaults",
        "x-suxen-required-privilege": "admin:download-gate-defaults:write",
        "summary": "Delete the instance-wide download-gate default",
        "tags": [
          "Policies"
        ],
        "responses": {
          "204": {
            "description": "Instance-wide download-gate default deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Local username and password."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token or OIDC JWT",
        "description": "Suxen API token or a configured OIDC bearer token. API-token scopes use repository:<name|*>:<read|write|delete|annotate|manage> and admin:<resource|*>:<action|*>. An empty scope list does not further restrict the user's roles.",
        "x-suxen-scope-syntax": [
          "repository:<name|*>:<read|write|delete|annotate|manage>",
          "admin:<resource|*>:<action|*>"
        ]
      },
      "oidcSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "suxen_oidc_token",
        "description": "HTTP-only browser session established by the configured OIDC login flow."
      }
    },
    "headers": {
      "Location": {
        "description": "Canonical path of the created resource.",
        "schema": {"type": "string"}
      }
    },
    "parameters": {
...
      "RepositoryName": {
        "name": "name",
        "in": "path",
        "required": true,
        "description": "Repository name",
        "schema": {
          "$ref": "#/components/schemas/ResourceName"
        }
      },
...
      "Force": {
        "name": "force",
        "in": "query",
        "required": false,
        "description": "Transfer a managed resource out of declarative provisioning before applying the mutation",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request. Account validation codes include invalid_username and password_required; the reserved repository name default returns reserved_repository_name.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required. Control-plane routes answer 401 without a WWW-Authenticate challenge so browsers do not raise the native Basic-auth popup; clients authenticate with a bearer token or the cookie session and treat 401 as a prompt to sign in.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient privilege",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "Resource already exists, is declaratively managed, or is still referenced",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Error": {
        "description": "Error response",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
      "ResourceName": {
        "type": "string",
        "maxLength": 255,
        "pattern": "^[a-z0-9_-]+$",
        "example": "releases"
      },
      "Timestamp": {
        "type": "string",
        "format": "date-time"
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 7807 problem details with a stable suxen error code.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "urn:suxen:problem:not_found"
          },
          "title": {
            "type": "string",
            "example": "Not found"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "pattern": "^[a-z0-9_]+$",
            "example": "not_found"
          }
        }
      },
...
      "DownloadGateRequest": {
        "type": "object",
        "required": ["criteria"],
        "properties": {
          "criteria": {"$ref": "#/components/schemas/Criteria"},
          "enabled": {
            "type": "boolean",
            "default": true
          },
          "inheritGlobal": {
            "type": "boolean",
            "default": true,
            "description": "When true, the instance-wide default gate's criteria AND-extend this repository's."
          }
        }
      },
      "DownloadGateUpdateRequest": {
        "type": "object",
        "description": "Replaces the repository gate. criteria must be a non-null array; [] clears local criteria, while inherited criteria still apply unless inheritance is disabled. Missing or null criteria are rejected without changing policy or ownership.",
        "additionalProperties": false,
        "required": ["criteria"],
        "properties": {
          "criteria": {"$ref": "#/components/schemas/CriteriaRequest"},
          "enabled": {"type": "boolean", "default": true},
          "inheritGlobal": {"type": "boolean", "default": true}
        }
      },
      "DownloadGateDefaultsUpdateRequest": {
        "type": "object",
        "description": "Replaces the instance-wide gate. criteria must be a non-null array; [] clears criteria. Missing or null criteria are rejected without changing policy or ownership.",
        "additionalProperties": false,
        "required": ["criteria"],
        "properties": {
          "criteria": {"$ref": "#/components/schemas/CriteriaRequest"},
          "enabled": {"type": "boolean", "default": true}
        }
      },
      "DownloadGate": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DownloadGateRequest"
          },
          {
            "type": "object",
            "required": [
              "repository",
              "managed",
              "updatedAt"
            ],
            "properties": {
              "repository": {
                "type": "string"
              },
              "managed": {
                "type": "boolean",
                "readOnly": true
              },
              "updatedAt": {
                "$ref": "#/components/schemas/Timestamp"
              }
            }
          }
        ]
      },
...
      "Predicate": {
        "type": "object",
        "description": "One condition over the projected asset attribute view. All predicates in a policy criterion list must match. The reserved roots are sys, classification, provenance, and every registered format name returned in discovery.formats (including raw and oci). Stable projected paths are listed in x-suxen-reserved-paths and discovery.attributePaths; format plugins may expose additional paths under their reserved roots. classification.<key> addresses rule-assigned labels with dynamic keys. Other roots address writer-owned attribute namespaces.",
        "required": [
          "path",
          "op"
        ],
        "properties": {
          "path": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+(\\.[A-Za-z0-9_-]+)*$",
            "example": "sys.lastAccessed"
          },
          "op": {
            "type": "string",
            "enum": [
              "=",
              "!=",
              "<",
              "<=",
              ">",
              ">=",
              "before",
              "after",
              "matches",
              "contains",
              "in",
              "not-in",
              "exists",
              "absent"
            ]
          },
          "value": {
            "description": "Operator literal. Omit for exists and absent. Use a non-empty array for in and not-in, an RE2 string for matches, and RFC3339 or a non-negative duration using Go, d, w, mo, or y units for before and after. JSON numbers compare by exact decimal value, including large exponents and numbers nested in objects or arrays."
          }
        },
        "x-suxen-reserved-paths": {
          "sys": [
            "repository",
            "format",
            "type",
            "blobStore",
            "path",
            "digest",
            "size",
            "contentType",
            "kind",
            "createdAt",
            "updatedAt",
            "lastAccessed",
            "validatedAt"
          ],
          "raw": [
            "path"
          ],
          "oci": [
            "image",
            "tag",
            "digest",
            "mediaType",
            "artifactType"
          ],
          "classification": [
            "<key>"
          ],
          "provenance": [
            "status",
            "format",
            "fingerprint",
            "identity",
            "issuer",
            "reason",
            "digest",
            "policyUpdatedAt",
            "verifiedAt"
          ]
        }
      },
      "Criteria": {
        "type": "array",
        "description": "Ordered, conjunctive policy predicates.",
        "minItems": 1,
        "items": {
          "$ref": "#/components/schemas/Predicate"
        }
      },
      "PredicateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["path", "op"],
        "properties": {
          "path": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+(\\.[A-Za-z0-9_-]+)*$"
          },
          "op": {
            "type": "string",
            "enum": [
              "=", "!=", "<", "<=", ">", ">=", "before", "after",
              "matches", "contains", "in", "not-in", "exists", "absent"
            ]
          },
          "value": {
            "description": "Operator literal. Omit for exists and absent."
          }
        }
      },
      "CriteriaRequest": {
        "type": "array",
        "description": "Ordered, conjunctive policy predicates.",
        "minItems": 1,
        "items": {"$ref": "#/components/schemas/PredicateRequest"}
      },
...
    }
  }
}